backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pillow
sentence-transformers
torch
scikit-learn
numpy
passlib[bcrypt]
bcrypt==4.0.1
alembic
pgvector
cloudinary
python-dotenv
slowapi
pyjwt
pydantic-settings
email-validator
